Geocode amenities

This function search amenities as defined by OpenStreetMap on a restricted area defined by a bounding box in the form of (<min_latitude>, <min_longitude>, <max_latitude>, <max_longitude>).

geo_amenity(
  bbox,
  amenity,
  lat = "lat",
  long = "lon",
  limit = 1,
  full_results = FALSE,
  return_addresses = TRUE,
  verbose = FALSE,
  custom_query = list(),
  strict = FALSE
)

Arguments

bbox
A numeric vector of latitude and longitude (<min_latitude>, <min_longitude>, <max_latitude>, <max_longitude>) that restrict the search area. See Details.

amenity
A character of a vector of character with the amenities to be geolocated (i.e. c("pub", "restaurant")). See Details or osm_amenities.

lat
latitude column name (i.e. "lat").

long
longitude column name (i.e. "long").

limit
maximum number of results to return per input address. Note that each query returns a maximum of 50 results.

full_results
returns all data from the geocoder service if TRUE. If FALSE then only longitude and latitude are returned from the geocoder service.

return_addresses
return input addresses with results if TRUE. Note that most services return the input addresses with full_results = TRUE and setting return_addresses to FALSE does not prevent this.

verbose
if TRUE then detailed logs are output to the console

custom_query
API-specific parameters to be used. See geo_lite().

strict
Logical TRUE/FALSE. Force the results to be included inside the bbox. Note that Nominatim default behaviour may return results located outside the provided bounding box.

Value

A tibble with the results.

Details

Bounding boxes can be located using different online tools, as Bounding Box Tool.

For a full list of valid amenities see https://wiki.openstreetmap.org/wiki/Key:amenity.

See also

Examples

# \donttest{
# Times Square, NY, USA
bbox <- c(
  -73.9894467311, 40.75573629,
  -73.9830630737, 40.75789245
)

geo_amenity(
  bbox = bbox,
  amenity = "restaurant"
)

#> # A tibble: 1 × 4
#>   query        lat   lon address                                                
#>   <chr>      <dbl> <dbl> <chr>                                                  
#> 1 restaurant  40.8 -74.0 Bond45, West 45th Street, Times Square, Manhattan, New…

# Several amenities
geo_amenity(
  bbox = bbox,
  amenity = c("restaurant", "pub")
)

#> # A tibble: 2 × 4
#>   query        lat   lon address                                                
#>   <chr>      <dbl> <dbl> <chr>                                                  
#> 1 restaurant  40.8 -74.0 Bond45, West 45th Street, Times Square, Manhattan, New…
#> 2 pub         40.8 -74.0 Connolly's, 121, West 45th Street, Times Square, Manha…

# Increase limit and use with strict
geo_amenity(
  bbox = bbox,
  amenity = c("restaurant", "pub"),
  limit = 10,
  strict = TRUE
)

#> # A tibble: 16 × 4
#>    query        lat   lon address                                               
#>    <chr>      <dbl> <dbl> <chr>                                                 
#>  1 restaurant  40.8 -74.0 Sardi's, 234, West 44th Street, Theater District, Man…
#>  2 restaurant  40.8 -74.0 Bond45, West 45th Street, Times Square, Manhattan, Ne…
#>  3 restaurant  40.8 -74.0 Chevys, 259, West 42nd Street, Theater District, Manh…
#>  4 restaurant  40.8 -74.0 Applebee's, 234, West 42nd Street, Theater District, …
#>  5 restaurant  40.8 -74.0 Planet Hollywood, 1540, Broadway, Times Square, Manha…
#>  6 restaurant  40.8 -74.0 Bubba Gump Shrimp Co., 1501, Broadway, Times Square, …
#>  7 restaurant  40.8 -74.0 Carmine's, 200, West 44th Street, Theater District, M…
#>  8 restaurant  40.8 -74.0 Dallas BBQ, 241, West 42nd Street, Theater District, …
#>  9 restaurant  40.8 -74.0 bella vita tranttoria, 211, West 43rd Street, Theater…
#> 10 restaurant  40.8 -74.0 Virgil's Real BBQ, 152, West 44th Street, Theater Dis…
#> 11 pub         40.8 -74.0 Connolly's, 121, West 45th Street, Times Square, Manh…
#> 12 pub         40.8 -74.0 Perfect Pint, 123, West 45th Street, Times Square, Ma…
#> 13 pub         40.8 -74.0 O'Lunney's, 145, West 45th Street, Times Square, Manh…
#> 14 pub         40.8 -74.0 Jimmy's Corner, 140, West 44th Street, Theater Distri…
#> 15 pub         40.8 -74.0 BXL Cafe, 125, West 43rd Street, Theater District, Ma…
#> 16 pub         40.8 -74.0 O’Brien’s, 134, West 46th Street, Midtown, Times Squa…

# }