Geocode amenities

This function search amenities as defined by OpenStreetMap on a restricted area defined by a bounding box in the form of (<min_latitude>, <min_longitude>, <max_latitude>, <max_longitude>). This function returns the tibble associated with the query, see geo_amenity_sf() for retrieving the data as a spatial object (sf format).

Usage

geo_amenity(
  bbox,
  amenity,
  lat = "lat",
  long = "lon",
  limit = 1,
  full_results = FALSE,
  return_addresses = TRUE,
  verbose = FALSE,
  custom_query = list(),
  strict = FALSE
)

Arguments

bbox

A numeric vector of latitude and longitude (<min_latitude>, <min_longitude>, <max_latitude>, <max_longitude>) that restrict the search area. See Details.

amenity

A character of a vector of character with the amenities to be geolocated (i.e. c("pub", "restaurant")). See Details and osm_amenities.

lat

latitude column name in the output data (default "lat").

long

longitude column name in the output data (default "long").

limit

maximum number of results to return per input address. Note that each query returns a maximum of 50 results.

full_results

returns all available data from the API service. If FALSE (default) only latitude, longitude and address columns are returned. See also return_addresses.

return_addresses

return input addresses with results if TRUE.

verbose

if TRUE then detailed logs are output to the console.

custom_query

API-specific parameters to be used. See geo_lite().

strict

Logical TRUE/FALSE. Force the results to be included inside the bbox. Note that Nominatim default behavior may return results located outside the provided bounding box.

Value

A tibble with the results.

Details

Bounding boxes can be located using different online tools, as Bounding Box Tool.

For a full list of valid amenities see https://wiki.openstreetmap.org/wiki/Key:amenity.

See also

Examples

# \donttest{
# Times Square, NY, USA
bbox <- c(-73.9894467311, 40.75573629, -73.9830630737, 40.75789245)

geo_amenity(bbox = bbox, amenity = "restaurant")
#> # A tibble: 1 × 4
#>   query        lat   lon address                                                
#>   <chr>      <dbl> <dbl> <chr>                                                  
#> 1 restaurant  40.8 -74.0 Sardi's, 234, West 44th Street, Theater District, Manh…

# Several amenities
geo_amenity(bbox = bbox, amenity = c("restaurant", "pub"))
#> # A tibble: 2 × 4
#>   query        lat   lon address                                                
#>   <chr>      <dbl> <dbl> <chr>                                                  
#> 1 restaurant  40.8 -74.0 Sardi's, 234, West 44th Street, Theater District, Manh…
#> 2 pub         40.8 -74.0 BXL Cafe, 125, West 43rd Street, Theater District, Man…

# Increase limit and use with strict
geo_amenity(
  bbox = bbox, amenity = c("restaurant", "pub"), limit = 10,
  strict = TRUE
)
#> # A tibble: 16 × 4
#>    query        lat   lon address                                               
#>    <chr>      <dbl> <dbl> <chr>                                                 
#>  1 restaurant  40.8 -74.0 Sardi's, 234, West 44th Street, Theater District, Man…
#>  2 restaurant  40.8 -74.0 Dave & Buster's, 234, West 42nd Street, Theater Distr…
#>  3 restaurant  40.8 -74.0 Applebee's, 234, West 42nd Street, Theater District, …
#>  4 restaurant  40.8 -74.0 Brooklyn Deli, 1501, Broadway, Theater District, Manh…
#>  5 restaurant  40.8 -74.0 Dallas BBQ, 241, West 42nd Street, Theater District, …
#>  6 restaurant  40.8 -74.0 Villa Fresh Italian Kitchen, 263, West 42nd Street, T…
#>  7 restaurant  40.8 -74.0 Carmine's, 200, West 44th Street, Theater District, M…
#>  8 restaurant  40.8 -74.0 The Lambs Club, West 44th Street, Theater District, M…
#>  9 restaurant  40.8 -74.0 Haru Sushi, 229, West 43rd Street, Theater District, …
#> 10 restaurant  40.8 -74.0 Hard Rock Cafe, 1501, Broadway, Theater District, Man…
#> 11 pub         40.8 -74.0 BXL Cafe, 125, West 43rd Street, Theater District, Ma…
#> 12 pub         40.8 -74.0 O'Lunney's, 145, West 45th Street, Times Square, Manh…
#> 13 pub         40.8 -74.0 Jimmy's Corner, 140, West 44th Street, Theater Distri…
#> 14 pub         40.8 -74.0 O’Brien’s, 134, West 46th Street, Times Square, Manha…
#> 15 pub         40.8 -74.0 Connolly's, 121, West 45th Street, Times Square, Manh…
#> 16 pub         40.8 -74.0 Perfect Pint, 123, West 45th Street, Times Square, Ma…
# }