Address Search API for OSM objects

Geocodes addresses given as character values. This function returns the tibble associated with the query, see geo_lite_sf() for retrieving the data as a spatial object (sf format).

Usage

geo_lite(
  address,
  lat = "lat",
  long = "lon",
  limit = 1,
  full_results = FALSE,
  return_addresses = TRUE,
  verbose = FALSE,
  custom_query = list()
)

Arguments

address

character with single line address ("1600 Pennsylvania Ave NW, Washington") or a vector of addresses (c("Madrid", "Barcelona")).

lat

latitude column name in the output data (default "lat").

long

longitude column name in the output data (default "long").

limit

maximum number of results to return per input address. Note that each query returns a maximum of 50 results.

full_results

returns all available data from the API service. If FALSE (default) only latitude, longitude and address columns are returned. See also return_addresses.

return_addresses

return input addresses with results if TRUE.

verbose

if TRUE then detailed logs are output to the console.

custom_query

A named list with API-specific parameters to be used (i.e. list(countrycodes = "US")). See Details.

Value

A tibble with the results.

Details

See https://nominatim.org/release-docs/latest/api/Search/ for additional parameters to be passed to custom_query.

See also

Examples

# \donttest{
geo_lite("Madrid, Spain")
#> # A tibble: 1 × 4
#>   query           lat   lon address                            
#>   <chr>         <dbl> <dbl> <chr>                              
#> 1 Madrid, Spain  40.4 -3.70 Madrid, Comunidad de Madrid, España

# Several addresses
geo_lite(c("Madrid", "Barcelona"))
#> # A tibble: 2 × 4
#>   query       lat   lon address                                                 
#>   <chr>     <dbl> <dbl> <chr>                                                   
#> 1 Madrid     40.4 -3.70 Madrid, Comunidad de Madrid, España                     
#> 2 Barcelona  41.4  2.18 Barcelona, Barcelonès, Barcelona, Catalunya, 08001, Esp…

# With options: restrict search to USA
geo_lite(c("Madrid", "Barcelona"),
  custom_query = list(countrycodes = "US"),
  full_results = TRUE
)
#> # A tibble: 2 × 24
#>   query       lat   lon address     place_id licence osm_type osm_id class type 
#>   <chr>     <dbl> <dbl> <chr>          <int> <chr>   <chr>     <int> <chr> <chr>
#> 1 Madrid     41.9 -93.8 Madrid, Bo… 28064051 Data ©… relation 1.29e5 boun… admi…
#> 2 Barcelona  42.3 -79.6 Barcelona,…  6486048 Data ©… node     1.58e8 place haml…
#> # ℹ 14 more variables: place_rank <int>, importance <dbl>, ype <chr>,
#> #   name <chr>, boundingbox <list>, village <chr>, county <chr>, state <chr>,
#> #   `ISO3166-2-lvl4` <chr>, country <chr>, country_code <chr>, hamlet <chr>,
#> #   town <chr>, postcode <chr>
# }