nominatimlite: Interface with 'Nominatim' API Service

Lite interface for getting data from 'OSM' service 'Nominatim' https://nominatim.org/release-docs/latest/. Extract coordinates from addresses, find places near a set of coordinates and return spatial objects on 'sf' format.

See also

Author

Maintainer: Diego Hernangómez diego.hernangomezherrero@gmail.com (ORCID) [copyright holder]

Other contributors:

Jindra Lacko (ORCID) [contributor, reviewer]

Alex White [contributor]

OpenStreetMap (For the data) [copyright holder]