OpenStreetMap amenity database

Database with the list of amenities available on OpenStreetMap.

Format

A tibble with the amenities and the corresponding category

Source

https://wiki.openstreetmap.org/wiki/Key:amenity

Details

category	amenity

Sustenance	bar

Sustenance	biergarten

Sustenance	cafe

Sustenance	fast_food

Sustenance	food_court

Sustenance	ice_cream

Sustenance	pub

Sustenance	restaurant

Education	college

Education	driving_school

Education	kindergarten

Education	language_school

Education	library

Education	toy_library

Education	music_school

Education	school

Education	university

Transportation	bicycle_parking

Transportation	bicycle_repair_station

Transportation	bicycle_rental

Transportation	boat_rental

Transportation	boat_sharing

Transportation	bus_station

Transportation	car_rental

Transportation	car_sharing

Transportation	car_wash

Transportation	vehicle_inspection

Transportation	charging_station

Transportation	ferry_terminal

Transportation	fuel

Transportation	grit_bin

Transportation	motorcycle_parking

Transportation	parking

Transportation	parking_entrance

Transportation	parking_space

Transportation	taxi

Financial	atm

Financial	bank

Financial	bureau_de_change

Healthcare	baby_hatch

Healthcare	clinic

Healthcare	dentist

Healthcare	doctors

Healthcare	hospital

Healthcare	nursing_home

Healthcare	pharmacy

Healthcare	social_facility

Healthcare	veterinary

Entertainment-Arts-Culture	arts_centre

Entertainment-Arts-Culture	brothel

Entertainment-Arts-Culture	casino

Entertainment-Arts-Culture	cinema

Entertainment-Arts-Culture	community_centre

Entertainment-Arts-Culture	conference_centre

Entertainment-Arts-Culture	events_venue

Entertainment-Arts-Culture	fountain

Entertainment-Arts-Culture	gambling

Entertainment-Arts-Culture	love_hotel

Entertainment-Arts-Culture	nightclub

Entertainment-Arts-Culture	planetarium

Entertainment-Arts-Culture	public_bookcase

Entertainment-Arts-Culture	social_centre

Entertainment-Arts-Culture	stripclub

Entertainment-Arts-Culture	studio

Entertainment-Arts-Culture	swingerclub

Entertainment-Arts-Culture	theatre

Public Service	courthouse

Public Service	embassy

Public Service	fire_station

Public Service	police

Public Service	post_box

Public Service	post_depot

Public Service	post_office

Public Service	prison

Public Service	ranger_station

Public Service	townhall

Facilities	bbq

Facilities	bench

Facilities	dog_toilet

Facilities	drinking_water

Facilities	give_box

Facilities	shelter

Facilities	shower

Facilities	telephone

Facilities	toilets

Facilities	water_point

Facilities	watering_place

Waste Management	sanitary_dump_station

Waste Management	recycling

Waste Management	waste_basket

Waste Management	waste_disposal

Waste Management	waste_transfer_station

Others	animal_boarding

Others	animal_breeding

Others	animal_shelter

Others	baking_oven

Others	childcare

Others	clock

Others	crematorium

Others	dive_centre

Note

Data extracted on 14 June 2021.

See also

Examples

amenities <- nominatimlite::osm_amenities

amenities

#> # A tibble: 100 × 2
#>    category   amenity       
#>    <chr>      <chr>         
#>  1 Sustenance bar           
#>  2 Sustenance biergarten    
#>  3 Sustenance cafe          
#>  4 Sustenance fast_food     
#>  5 Sustenance food_court    
#>  6 Sustenance ice_cream     
#>  7 Sustenance pub           
#>  8 Sustenance restaurant    
#>  9 Education  college       
#> 10 Education  driving_school
#> # … with 90 more rows